CPDTP-555 Remove evidence held for NPQ from request body
