+++ app/services/record_declarations/retained/ecf.rb
@@ -2,13 +2,33 @@
 
 module RecordDeclarations
   module Retained
-    class ECF < Base
+    class ECF < ::RecordDeclarations::Base
       include RecordDeclarations::ECF
 
       class << self
         def valid_evidence_types
           %w[training-event-attended self-study-material-completed other].freeze
         end
+      end
+
+      attr_accessor :evidence_held
+
+      validates :evidence_held, presence: { message: I18n.t(:missing_evidence_held) }
+      validates :evidence_held, inclusion: { in: :valid_evidence_types, message: I18n.t(:invalid_evidence_type) }, allow_blank: true
+
+    private
+
+      def initialize(params:)
+        super(params: params)
+        self.evidence_held = params[:evidence_held]
+      end
+
+      def valid_evidence_types
+        self.class.valid_evidence_types
+      end
+
+      def declaration_parameters
+        super.merge(evidence_held: evidence_held)
       end
     end
   end
