+++ etc/schema/0.3/npq/participant_declarations/retained/create/request.json
@@ -5,8 +5,7 @@
       "declaration_type": "retained-1",
       "declaration_date": "2015-02-14T02:21:34",
       "participant_id": "0cdc3c2f-20e3-4721-97f1-485b397ae7e9",
-      "course_type": "leading-teaching",
-      "evidence_held": "yes"
+      "course_type": "leading-teaching"
     }
   }
 }
