+++ spec/services/record_declarations/retained/early_career_teacher_spec.rb
@@ -17,7 +17,7 @@ RSpec.describe RecordDeclarations::Retained::EarlyCareerTeacher do
     travel_to cutoff_start_datetime + 2.days
   end
 
-  it_behaves_like "a retained participant declaration service" do
+  it_behaves_like "a participant declaration with evidence held service" do
     def given_params
       retained_ect_params
     end
