+++ spec/services/record_declarations/retained/npq_spec.rb
@@ -16,7 +16,7 @@ RSpec.describe RecordDeclarations::Retained::NPQ do
     travel_to cutoff_start_datetime + 2.days
   end
 
-  it_behaves_like "a retained participant declaration service" do
+  it_behaves_like "a participant declaration without evidence held service" do
     def given_params
       retained_npq_params
     end
