+++ spec/services/record_declarations/started/early_career_teacher_spec.rb
@@ -16,7 +16,7 @@ RSpec.describe RecordDeclarations::Started::EarlyCareerTeacher do
     travel_to cutoff_start_datetime + 2.days
   end
 
-  it_behaves_like "a started participant declaration service" do
+  it_behaves_like "a participant declaration without evidence held service" do
     def given_params
       ect_params
     end
